refactor(mywebscript): improve error handling and display
Changes made refine error handling on the server-side to better align with common client error standards and boost error visibility in the interface. Now, the HTTP status code for data invalidity errors is set to 400, while adjustments made to the user interface provide a cleaner, more aesthetic display of errors.